+++ R/to_any_case.R
@@ -71,7 +71,7 @@
 #'  of length 2, the second letter will be transliterated to lowercase, for example Oe, Ae, Ss, which
 #'  might not always be what is intended. In this case you can make usage of the option to supply named elements and specify the transliterations yourself.
 #' 
-#' @param numerals A character specifying the alignment of numerals (\code{"middle"}, \code{left}, \code{right} or \code{asis}). I.e. \code{numerals = "left"} ensures that no output separator is in front of a digit.
+#' @param numerals A character specifying the alignment of numerals (\code{"middle"}, \code{left}, \code{right}, \code{asis} or \code{tight}). I.e. \code{numerals = "left"} ensures that no output separator is in front of a digit.
 #' 
 #' @param sep_out (short for separator output) String that will be used as separator. The defaults are \code{"_"} 
 #' and \code{""}, regarding the specified \code{case}.
@@ -166,7 +166,7 @@ to_any_case <- function(string,
                         sep_in = "[^[:alnum:]]",
                         parsing_option = 1,
                         transliterations = NULL,
-                        numerals = c("middle", "left", "right", "asis"),
+                        numerals = c("middle", "left", "right", "asis", "tight"),
                         sep_out = NULL,
                         unique_sep = NULL,
                         empty_fill = NULL,
@@ -339,12 +339,12 @@ if (case != "swap") {
     if (parsing_option == 4 | numerals == "asis") {
       string <- stringr::str_replace_all(string, " ", "")
     }
-    if (numerals == "right") {
+    if (numerals == "right" | numerals == "tight") {
       # underscore with a digit before and no digit after
       string <- stringr::str_replace_all(string, "(?<=\\d)_(?!\\d)", "")
     }
     
-    if (numerals == "left") {
+    if (numerals == "left" | numerals == "tight") {
       # underscore with no digit before and a digit after
       string <- stringr::str_replace_all(string, "(?<!\\d)_(?=\\d)", "")
     }