+++ NEWS.md
@@ -1,3 +1,7 @@
+# snakecase 0.9.2.9001
+
+* **numerals**: new alignment option `"tight"` which allows to suppress all underscores between numerals and letters.
+
 # snakecase 0.9.2
 
 * **cases**: added `to_sentence_case()` (same as snake, but first letter is uppercase and default sep_out is space).